Removed all emojis and mentions from SimpleQuill. Once again, it's looking like Quill is just not the solution to editor. Ugh.
