+++ cyfr-web/components/ui/MentionsMenu.tsx
@@ -19,8 +19,74 @@ const MentionsMenu = ({onSelect, searchTerm='', index, setIndex, showMenu, setSh
     const [search, setSearch] = useState<string>('')
     const [mentions, setMentions] = useState<User[]>([])
     const [isLoading, setIsLoading] = useState(true)
+    const [showMentions, setShowMentions] = useState(false)
+    const [mentionSearch, setMentionSearch] = useState<string>('')
+    const [mentionIndex, setMentionIndex] = useState<number>(0)
     
     const repl = (id:string, name:string) => `<span class="mention-link" username="${name}" userid="${id}">@${name}</span>`
+
+    const ignoreKey = (e:KeyboardEvent) => {
+        const r = ['Control', 'Delete', 'Shift', `Escape`, `CapsLock`,`Backspace`,`#$%^&*();:"'<>,./?\\`].indexOf(e.key)
+        log(`IgnoreKey: ${e.key} : ${r}`)
+        return r >= 0
+      }
+      const cancelKey = (e:KeyboardEvent) => {
+        const r = ['Escape'].indexOf(e.key)
+        return r >= 0
+      }
+      const selectKey = (e:KeyboardEvent) => {
+        const r = ['Tab', 'Space', 'Enter'].indexOf(e.key)
+        return r >= 0
+      }
+      const alpha = (e:KeyboardEvent) => ignoreKey(e) !== false && RegExp(`^[A-Za-z0-9_@.+-]*$`).test(e.code) === true
+    
+      const handleKey = (e:KeyboardEvent) => {
+        const closeSearch = () => {
+          setShowMentions(() => false)
+          setMentionSearch((s) => '')
+        }
+    
+        if(e.key==='@') {
+          log(`Squill.handleKey @`)
+          setShowMentions(() => true)
+        } else if (!showMentions) {
+          return
+        }
+    
+        e.preventDefault()
+    
+        if (ignoreKey(e)) {
+          log(`ignoring ${e.key}`)
+        }
+        else if(cancelKey(e)) {
+          log(`Squill.cancelKey`)
+          e.preventDefault()
+          closeSearch()
+        }
+        else if (selectKey(e)) {
+          log(`Squill.acceptKey`)
+          closeSearch()
+        }
+        else if(e.key==='ArrowDown') {
+          log(`Squill.handleKey ArrowDown`)
+          setMentionIndex((i) => i+1)
+        }
+        else if(e.key==='ArrowUp') {
+          log(`Squill.handleKey ArrowUp`)
+          setMentionIndex((i) => i <= 0 ? 0 : i-1)
+        }
+        else if(e.key==='Backspace') {
+          log(`Squill.handleKey Backspace`)
+          setMentionSearch((s) => s.length > 0 ? s.substring(0,s.length-1) : '')
+          if (mentionSearch.length<1) {
+            closeSearch()
+          }
+        }
+        else if (showMentions) {
+          log(`sending to mentions ${e.key}`)
+          setMentionSearch((s) => s+e.key)
+        }
+      }
 
     const onChoose = (user:User) => {
         const chose = repl(user.id, (user.name||user.email)!)
