+++ cyfr-web/package.json
@@ -46,6 +46,7 @@
     "nodemailer": "^6.8.0",
     "prisma": "^4.9.0",
     "quill": "^1.3.7",
+    "quill-emoji": "^0.2.0",
     "quill-mention": "^3.1.0",
     "react": "^18.2.0",
     "react-debounce-input": "^3.3.0",
