+++ cyfr-web/quill.d.ts
@@ -1,0 +1,2 @@
+declare module 'quill-mention'
+declare module 'quill-emoji'
